+++ app/src/main/java/dz/deepwork/gamemn02/presenceregister/data/sessions/SessionDao.java
@@ -11,5 +11,5 @@ public interface SessionDao {
     long[] insert(Session... sessions);
 
     @Query("SELECT * FROM sessions WHERE member_Id = :memberId")
-    Session find(int memberId);
+    Session find(long memberId);
 }
